+++ src/main/java/com/greenfoxacademy/springwebapp/product/models/ProductRequestDTO.java
@@ -5,8 +5,8 @@ import javax.validation.constraints.NotBlank;
 public class ProductRequestDTO {
   private String name;
   private String quality;
-  private Float size;
-  private Float length;
+  private Double size;
+  private Double length;
   private Integer quantity;
 
 }
